+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 ## Not released
 - Add meta value example to List component story [#31](https://github.com/CartoDB/carto-react-lib/pull/31)
+- Mobile fixes and fix loading datasets for free accounts [#38](https://github.com/CartoDB/carto-react-lib/pull/38)
 - Add updateLayer action to cartoSlice [#39](https://github.com/CartoDB/carto-react-lib/pull/39)
 
 ## 1.0.0-beta6 (2020-11-27)